+++ src/content-scripts/components/InviteToOpenSauced/InviteToOpenSaucedModal.ts
@@ -18,7 +18,7 @@ export const InviteToOpenSaucedModal = (
 ) => {
     const emailBody =
     typeof emailAddress === "string" &&
-    `Hey ${username}. I'm using OpenSauced to keep track of my contributions and discover new projects. Check it out at https://opensauced.pizza/`;
+    `Hey ${username}. I'm using OpenSauced to keep track of my contributions and discover new projects. Try connecting your GitHub to https://opensauced.pizza/`;
     const emailHref =
     typeof emailAddress === "string" &&
     `mailto:${emailAddress}?subject=${encodeURIComponent(