+++ src/popup/pages/home.tsx
@@ -1,6 +1,4 @@
 import {
-    HiArrowLeft,
-    HiArrowRight,
     HiArrowTopRightOnSquare,
     HiOutlineQuestionMarkCircle,
     HiPencil,
@@ -8,16 +6,17 @@ import {
 } from "react-icons/hi2";
 import { IoLogoLinkedin } from "react-icons/io5";
 import { FiSettings } from "react-icons/fi";
+import { Navigation, Pagination, A11y } from "swiper";
+import { Swiper, SwiperSlide } from "swiper/react";
+import "swiper/swiper-bundle.min.css";
 import OpenSaucedLogo from "../../assets/opensauced-logo.svg";
 import { useAuth } from "../../hooks/useAuth";
 import { useOpensaucedUserCheck } from "../../hooks/useOpensaucedUserCheck";
 import { Profile } from "./profile";
 import { goTo } from "react-chrome-extension-router";
-import AIPRDescription from "./aiprdescription";
 import PostOnHighlight from "./posthighlight";
-import { getHighlights } from "../../utils/fetchOpenSaucedApiData";
 import { getRepoAPIURL } from "../../utils/urlMatchers";
-
+import { getEmojis, getHighlights } from "../../utils/fetchOpenSaucedApiData";
 
 import Help from "./help";
 import { useEffect, useState } from "react";
@@ -25,13 +24,14 @@ import Settings from "./settings";
 import { OPEN_SAUCED_INSIGHTS_DOMAIN } from "../../constants";
 import type { Highlight } from "../../ts/types";
 import { usGetGitHubPageInfo } from "../../hooks/useGetGitHubPageInfo";
+import { HighlightSlide } from "../components/HighlightSlide";
 
 const Home = () => {
     const { user } = useAuth();
     const { currentTabIsOpensaucedUser, checkedUser } = useOpensaucedUserCheck();
     const { prUrl: pageURL, prTitle, type: GitHubPageType } = usGetGitHubPageInfo();
     const [highlights, setHighlights] = useState<Highlight[]>([]);
-    const [currentPage, setCurrentPage] = useState(0);
+    const [emojis, setEmojis] = useState<Record<string, string>[]>([]);
 
     useEffect(() => {
         const fetchHighlights = async () => {
@@ -49,16 +49,24 @@ const Home = () => {
             }
         };
 
+        const fetchEmojis = async () => {
+            try {
+                const emojiResponse = await getEmojis();
+                const emojis = emojiResponse.data;
+
+                if (!emojis) {
+                    return;
+                }
+
+                setEmojis(emojis);
+            } catch (error) {
+                console.log(error);
+            }
+        };
+
         void fetchHighlights();
+        void fetchEmojis();
     }, []);
-
-    const handlePrevious = () => {
-        setCurrentPage(prevPage => prevPage - 1);
-    };
-
-    const handleNext = () => {
-        setCurrentPage(prevPage => prevPage + 1);
-    };
 
     return (
         <div className="p-4 bg-slate-800">
@@ -66,8 +74,17 @@ const Home = () => {
                 <header className="flex justify-between">
                     <img
                         alt="OpenSauced logo"
-                        className="w-[45%]"
+                        className="w-[45%] cursor-pointer hover:opacity-80 transition-opacity duration-300 ease-in-out"
+                        role="presentation"
                         src={OpenSaucedLogo}
+                        onClick={() => {
+                            window.open(`https://${OPEN_SAUCED_INSIGHTS_DOMAIN}/feed`, "_blank");
+                        }}
+                        onKeyDown={e => {
+                            if (e.key === "Enter") {
+                                window.open(`https://${OPEN_SAUCED_INSIGHTS_DOMAIN}/feed`, "_blank");
+                            }
+                        }}
                     />
 
 
@@ -90,95 +107,41 @@ const Home = () => {
                 </header>
 
                 <main className="main-content">
-                    <h3 className="text font-medium text-base leading-10">Latest Highlight:</h3>
+                    <a
+                        className="flex items-center text-white hover:text-orange no-underline gap-2 w-full font-medium text-lg leading-10"
+                        href={`https://${OPEN_SAUCED_INSIGHTS_DOMAIN}/feed`}
+                        rel="noreferrer"
+                        target="_blank"
+                    >
+                            Highlights feed
+                        <HiArrowTopRightOnSquare />
+                    </a>
 
-                    {highlights.length > 0 && (
+                    {highlights.length > 0 && emojis.length > 0 && (
+                        <Swiper
+                            navigation
+                            modules={[Navigation, Pagination, A11y]}
+                            pagination={{ clickable: true }}
+                            slidesPerView={1}
+                            spaceBetween={50}
+                        >
+                            {
+                                highlights.map((highlight, index) => (
+                                    <SwiperSlide key={index}>
+                                        <HighlightSlide
+                                            emojis={emojis}
+                                            highlight={highlight}
+                                        />
+                                    </SwiperSlide>
+                                ))
+                            }
+                        </Swiper>
+                    )}
 
-                        <div className="border border-white/40 rounded-sm p-3 mt-2">
-                            <h3 className="text-base font-medium">
-                                <a
-                                    className="text-blue-500 hover:text-blue-700 underline cursor-pointer"
-                                    href={highlights[currentPage]?.url}
-                                    rel="noopener noreferrer"
-                                    target="_blank"
-                                >
-                                    {highlights[currentPage]?.title}
-                                </a>
-                            </h3>
-
-
-                            <div className="flex items-center">
-                                <div className="mr-2">Author:</div>
-
-                                <a
-                                    className="text-blue-500 hover:text-blue-700 underline cursor-pointer"
-                                    href={`https://insights.opensauced.pizza/user/${highlights[currentPage]?.login}`}
-                                    rel="noopener noreferrer"
-                                    target="_blank"
-
-
-                                >
-                                    {highlights[currentPage]?.login}
-                                </a>
-                            </div>
-
-                            <p className="py-2">
-                                {highlights[currentPage]?.highlight}
-                            </p>
-
-                            <div className="flex justify-center">
-                                <div className="flex justify-center mt-4">
-                                    <button
-                                        className="px-4 py-2 rounded-md bg-blue-500 text-white disabled:opacity-50"
-                                        disabled={currentPage === 0}
-                                        onClick={handlePrevious}
-                                    >
-                                        <HiArrowLeft />
-                                    </button>
-
-                                    <button
-                                        className="px-4 py-2 rounded-md bg-blue-500 text-white disabled:opacity-50 ml-4"
-                                        disabled={currentPage === highlights.length - 1}
-                                        onClick={handleNext}
-                                    >
-                                        <HiArrowRight />
-                                    </button>
-                                </div>
-                            </div>
-                        </div>)}
 
                     <h3 className="text font-medium text-base leading-10">Tools:</h3>
 
                     <div className="tools flex flex-col gap-2">
-                        <a
-                            className="flex items-center bg-slate-700 hover:bg-slate-700/70 text-white hover:text-orange no-underline gap-2 p-1.5 px-3 w-full rounded-sm font-medium text-sm"
-                            href={`https://${OPEN_SAUCED_INSIGHTS_DOMAIN}/feed`}
-                            rel="noreferrer"
-                            target="_blank"
-                        >
-                            <HiArrowTopRightOnSquare />
-                            Highlights feed
-                        </a>
-
-                        <a
-                            className="flex items-center bg-slate-700 hover:bg-slate-700/70 hover:text-orange text-white gap-2 p-1.5 px-3 w-full rounded-sm font-medium text-sm"
-                            href={`https://${OPEN_SAUCED_INSIGHTS_DOMAIN}/feed`}
-                            rel="noreferrer"
-                            target="_blank"
-                        >
-                            <HiArrowTopRightOnSquare />
-                            Dashboard
-                        </a>
-
-                        <button
-                            className="flex items-center bg-slate-700 hover:bg-slate-700/70 hover:text-orange text-white gap-2 p-1.5 px-3 w-full rounded-sm font-medium text-sm"
-                            onClick={() => {
-                                goTo(AIPRDescription);
-                            }}
-                        >
-                            <HiPencil />
-                            AI Configuration
-                        </button>
 
                         {GitHubPageType === "REPO" && (
                             <button